requirements.txt
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
python-dotenv
sentence-transformers
pypdf